Use owner name as site title for User and Organization sites.
Using the repo name as the site title makes sense for Projects, which can choose their own repo names. But User and Organization sites must use repo names like "{owner}.github.io" which:

* Does not look good as a browser tab title, bookmark, or header for a page that is supposed to be about a person or organization.
* Conveys no useful information, as it is unlikely that anyone would clone or fork such a repo.
* Looks confusing for users or organizations that use custom domains (CNAMEs), and have no other content indicating their site is hosted by GitHub.

The display name of the user or organization is a better choice. This is an optional property, so fall back to (login) name if display name is not present.

Resolve jekyll/github-metadata#196