Inject site.github via :pre_render step rather than :after_init
Jekyll 4 has a call to site.config.inspect, which renders all the values in site.config['github'].
This causes all of the API calls to be resolved regardless of whether the user is using them.
This patch should allow Jekyll 4 to call site.config.inspect without causing the large number of API calls.