Prevent loading all GH data on reset
The way this was setup, when running `build` or `serve` with `--watch`,
`inspect` was called and was forcing everything to load. This could
cause a very large number of API calls to GitHub. We now uninject the
MetadataDrop after rendering is complete. With this change
`GitHubMetadata` will no longer trigger all of those calls on reset.

The other scenario where a bunch of calls were being forced to be made
was when putting the `github` key with any sub-keys in `_config.yml`.
This went through every key in `GitHubMetadata` and loaded it, once
again potentially making a very large number of calls. These changes
take care of this scenario as well.